refactored to index
