+++ client/src/Components/LandingPage.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { ConnectSpotify } from './index';
+
+const LandingPage = () => <ConnectSpotify />;
+
+export default LandingPage;
