+++ client/src/Components/Navbar.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Typography from '@material-ui/core/Typography';
+
+const Navbar = () => (
+	<div>
+		<nav>
+			<Typography variant="h1" gutterBottom color="inherit">
+				MOODIFY
+			</Typography>
+		</nav>
+	</div>
+);
+
+export default Navbar;
