+++ client/src/Components/index.js
@@ -1,0 +1,24 @@
+// export { default as Navbar } from './navbar';
+// export { default as UserHome } from './user-home';
+// export { default as SingleBook } from './book-views/single-book';
+// export { default as AllBooks } from './book-views/all-books-view';
+// export { default as BooksByGenre } from './book-views/books-by-genre-view';
+// export { Login, Signup } from './auth-form';
+// export { default as LandingPage } from './landing-page';
+// export { default as Cart } from './payment/cart';
+// export { default as Checkout } from './payment/checkout';
+// export { default as StripeCheckout } from './payment/stripe-checkout';
+// export { default as ReviewOrder } from './payment/review-order';
+// export { default as AllStaffs } from './staff-views/all-staff';
+// export { default as SingleStaff } from './staff-views/single-staff';
+// export { default as Footer } from './footer';
+
+export { default as LandingPage } from './LandingPage';
+export { default as MainDisplay } from './MainDisplay';
+export { default as Navbar } from './Navbar';
+
+export { default as ConnectSpotify } from './Auth/ConnectSpotify';
+
+export { default as RecommendationsButton } from './RecommendationsButton';
+export { default as TrackComponent } from './TrackComponent';
+export { default as TrackContainer } from './TrackContainer';
