+++ client/src/index.js
@@ -5,7 +5,3 @@ import App from './App';
 require('dotenv').config();
 
 ReactDOM.render(<App />, document.getElementById('root'));
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
