--- client/src/navbar.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import Typography from '@material-ui/core/Typography';
-
-const Navbar = () => (
-  <div>
-    <nav>
-      <Typography variant="h1" gutterBottom color="inherit">
-        MOODIFY
-      </Typography>
-    </nav>
-  </div>
-);
-
-export default Navbar;
