+++ client/tests/RecommendationsButton.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import SpotifyContainer from './spotify-container';
+import RecommendationsButton from '../src/SpotifyComponents/RecommendationsButton';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<SpotifyContainer />, div);
+  ReactDOM.render(<RecommendationsButton />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
